add abstract player in expectance of draft players
